mobi-smart-transport-ai/services/public_data/requirements.txt
httpx==0.28.1
pydantic==2.10.3
python-dotenv==1.0.1
